Move DummyRequestStatement to a new module
