+++ src/nextline_schedule/scheduler.py
@@ -1,4 +1,3 @@
-import random
 from datetime import datetime, timedelta
 from logging import getLogger
 
@@ -36,23 +35,3 @@ class RequestStatement:
             raise RuntimeError(msg)
         self._logger.info(f'Response: {response.json()!r}')
         return response.json()['commands']
-
-
-class DummyRequestStatement:
-    async def __call__(self) -> str:
-        return generate_dummy_statement()
-
-
-def generate_dummy_statement() -> str:
-    s = '\n'.join(
-        f'time.sleep({random.uniform(1, 3):.2})' for _ in range(random.randint(5, 10))
-    )
-    s = '\n'.join(
-        (
-            f'# {datetime.utcnow().isoformat()}',
-            'import time',
-            '',
-            s,
-        )
-    )
-    return s
